Rework image caching

+++ src/Service/ImageCacheService.php
@@ -2,8 +2,10 @@
 
 namespace Movary\Service;
 
+use Doctrine\DBAL\Connection;
 use GuzzleHttp\Psr7\Request;
 use Movary\Util\File;
+use Movary\ValueObject\ResourceType;
 use Movary\ValueObject\Url;
 use Psr\Http\Client\ClientInterface;
 use Psr\Log\LoggerInterface;
@@ -17,6 +19,7 @@ class ImageCacheService
         private readonly File $fileUtil,
         private readonly LoggerInterface $logger,
         private readonly ClientInterface $httpClient,
+        private readonly Connection $dbConnection,
         private readonly string $publicDirectory,
         private readonly string $imageBasePath,
     ) {
@@ -26,9 +29,9 @@ class ImageCacheService
     /**
      * @return string|null Public path of image if file path was not already used
      */
-    public function cacheImage(Url $imageUrl, bool $forceRefresh = false) : ?string
+    public function cacheImage(Url $imageUrl, int $resourceId, ResourceType $resourceType, bool $forceRefresh) : ?string
     {
-        $imageFile = $this->publicDirectory . trim($this->imageBasePath, '/') . '/' . trim((string)$imageUrl->getPath(), '/');
+        $imageFile = $this->generateStorageFilePath($resourceId, $resourceType, $imageUrl);
 
         if ($forceRefresh === false && $this->fileUtil->fileExists($imageFile) === true) {
             return null;
@@ -51,19 +54,43 @@ class ImageCacheService
         $this->fileUtil->createFile($imageFile, $response->getBody()->getContents());
         $this->logger->debug('Cached image: ' . $imageUrl);
 
-        return $this->imageBasePath . trim((string)$imageUrl->getPath(), '/');
+        return str_replace($this->publicDirectory, '/', $imageFile);
     }
 
-    public function deleteImage(string $posterPath) : void
+    public function deleteImageByPosterPath(string $posterPath) : void
     {
         $imageFile = $this->publicDirectory . trim($posterPath, '/');
+
+        if ($this->fileUtil->fileExists($imageFile) === false) {
+            return;
+        }
 
         $this->fileUtil->deleteFile($imageFile);
     }
 
-    public function deleteImages() : void
+    public function deleteImagesByResourceType(ResourceType $resourceType) : void
     {
-        $this->fileUtil->deleteDirectoryContent($this->publicDirectory);
+        $this->fileUtil->deleteDirectoryContent($this->generateStorageDirectory($resourceType));
+    }
+
+    public function deleteOutdatedImagesByResourceType(ResourceType $resourceType) : void
+    {
+        $iterator = new \DirectoryIterator($this->generateStorageDirectory($resourceType));
+
+        foreach ($iterator as $file) {
+            if ($file->isDir() === true) {
+                continue;
+            }
+            $resourceId = pathinfo($file->getPathname(), PATHINFO_FILENAME);
+
+            $result = $this->dbConnection->executeQuery("SELECT id FROM $resourceType WHERE id = ?", [$resourceId]);
+
+            if ($result->fetchOne() !== false) {
+                continue;
+            }
+
+            $this->fileUtil->deleteFile($file->getPathname());
+        }
     }
 
     public function posterPathExists(string $posterPath) : bool
@@ -71,5 +98,30 @@ class ImageCacheService
         $imageFile = $this->publicDirectory . trim($posterPath, '/');
 
         return $this->fileUtil->fileExists($imageFile);
+    }
+
+    private function generateStorageDirectory(ResourceType $resourceType) : string
+    {
+        $pathId = match (true) {
+            $resourceType->isMovie() => 'movie',
+            $resourceType->isPerson() => 'person',
+
+            default => throw new \RuntimeException('Not handled resource type: ' . $resourceType)
+        };
+
+        return $this->publicDirectory . trim($this->imageBasePath, '/') . '/' . $pathId . '/';
+    }
+
+    private function generateStorageFilePath(int $resourceId, ResourceType $resourceType, Url $imageUrl) : string
+    {
+        $imageUrlPath = $imageUrl->getPath();
+
+        if ($imageUrlPath == null) {
+            throw new \RuntimeException('Could not get url path from image url: ' . $imageUrl);
+        }
+
+        $imageFileExtension = pathinfo($imageUrlPath, PATHINFO_EXTENSION);
+
+        return $this->generateStorageDirectory($resourceType) . $resourceId . '.' . $imageFileExtension;
     }
 }
